+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file, and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Added the context manager `pybryt.no_tracing`
+
 ## 0.1.0 - 2021-05-27
 
 * Added time complexity annotations and checking